requirements.txt
faker
python-dotenv
openai
langchain
langchain-openai
langchain-core
langchain-community
langchain-qdrant
qdrant-client
tiktoken
fastapi
uvicorn
streamlit
pydantic
